Fix resource leaks and memory corruptions.

+++ src/fragments.c
@@ -226,7 +226,7 @@ int delFrag(unsigned char * smac, int sequence)
 
 unsigned char * getCompleteFrag(unsigned char * smac,
 								int sequence,
-								int * packetlen,
+								size_t * packetlen,
 								int crypt,
 								unsigned char * wepkey,
 								int weplen)
@@ -299,7 +299,7 @@ unsigned char * getCompleteFrag(unsigned char * smac,
 
 					// delete captured fragments
 					delFrag(smac, sequence);
-					*packetlen = len;
+					*packetlen = (size_t) len;
 					return (packet);
 				}
 				else
@@ -322,7 +322,7 @@ unsigned char * getCompleteFrag(unsigned char * smac,
 
 				// delete captured fragments
 				delFrag(smac, sequence);
-				*packetlen = len;
+				*packetlen = (size_t) len;
 				return (packet);
 			}
 		}
